+++ agentic_doc/parse.py
@@ -16,7 +16,6 @@ from pydantic_core import Url
 from tqdm import tqdm
 import jsonschema
 from pypdf import PdfReader
-from dicttoxml import dicttoxml  # type: ignore[import-untyped]
 import xmltodict  # type: ignore[import-untyped]
 
 from agentic_doc.common import (
@@ -39,6 +38,8 @@ from agentic_doc.utils import (
     log_retry_failure,
     save_groundings_as_images,
     split_pdf,
+    _convert_pydantic_to_xml,
+    _fix_xml_dict,
 )
 
 _LOGGER = structlog.getLogger(__name__)
@@ -203,34 +204,6 @@ def _get_documents_from_bytes(doc_bytes: bytes) -> List[Path]:
     return [temp_file_path]
 
 
-def _fix_xml_dict(data: Any, parent_key: Optional[str] = None) -> Any:
-    """Convert XML dict structure back to original JSON structure"""
-    if isinstance(data, dict):
-        # Handle xmltodict's list conversion - when XML has multiple items with same tag
-        if "item" in data and isinstance(data["item"], list):
-            return [_fix_xml_dict(item) for item in data["item"]]
-        elif "item" in data:  # Single item case
-            return [_fix_xml_dict(data["item"])]
-        else:
-            result = {}
-            for k, v in data.items():
-                if k == "@type" or k.startswith("@"):  # Skip XML attributes
-                    continue
-                result[k] = _fix_xml_dict(v, k)
-            return result
-    elif isinstance(data, list):
-        return [_fix_xml_dict(item) for item in data]
-    elif data is None:
-        # Handle fields that should be empty lists when None
-        if parent_key in ["errors", "grounding", "chunks"]:
-            return []
-        elif parent_key in ["extraction", "extraction_metadata"]:
-            return {}
-        return None
-    else:
-        return data
-
-
 def _convert_to_parsed_documents(
     parse_results: Union[List[ParsedDocument[T]], List[Path]],
     result_save_dir: Optional[Union[str, Path]],
@@ -248,7 +221,7 @@ def _convert_to_parsed_documents(
             if config and config.output_xml:
                 with open(result, "r") as f:
                     data = xmltodict.parse(f.read())
-                    data = _fix_xml_dict(data["root"])
+                    data = _fix_xml_dict(data["document"])
             else:
                 with open(result, encoding="utf-8") as f:
                     data = json.load(f)
@@ -524,9 +497,7 @@ def parse_and_save_document(
         result_save_dir.mkdir(parents=True, exist_ok=True)
         if config and config.output_xml:
             save_path = result_save_dir / f"{result_name}.xml"
-            xml_bytes = dicttoxml(
-                json.loads(dump_parsed_doc_json(result)), attr_type=False
-            )
+            xml_bytes = _convert_pydantic_to_xml(result)
             save_path.write_text(xml_bytes.decode("utf-8"))
         else:
             save_path = result_save_dir / f"{result_name}.json"
